# CLN/PERF: Simplify argmin/argmax