--- a/asv_bench/asv.conf.json
+++ b/asv_bench/asv.conf.json
@@ -42,7 +42,6 @@
     // followed by the pip installed packages).
     "matrix": {
         "Cython": ["3.0"],
-        "build": [],
         "matplotlib": [],
         "sqlalchemy": [],
         "scipy": [],