--- a/pandas/core/nanops.py
+++ b/pandas/core/nanops.py
@@ -1439,20 +1439,15 @@ def _maybe_arg_null_out(
         return result
 
     if axis is None or not getattr(result, "ndim", False):
-        if skipna:
-            if mask.all():
-                raise ValueError("Encountered all NA values")
-        else:
-            if mask.any():
-                raise ValueError("Encountered an NA value with skipna=False")
-    else:
-        na_mask = mask.all(axis)
-        if skipna and na_mask.any():
+        if skipna and mask.all():
             raise ValueError("Encountered all NA values")
-        elif not skipna:
-            na_mask = mask.any(axis)
-            if na_mask.any():
-                raise ValueError("Encountered an NA value with skipna=False")
+        elif not skipna and mask.any():
+            raise ValueError("Encountered an NA value with skipna=False")
+    else:
+        if skipna and mask.all(axis).any():
+            raise ValueError("Encountered all NA values")
+        elif not skipna and mask.any(axis).any():
+            raise ValueError("Encountered an NA value with skipna=False")
     return result