- [ ] closes #xxxx (Replace xxxx with the GitHub issue number)
- [ ] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [ ] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [ ] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [ ] Added an entry in the latest `doc/source/whatsnew/vX.X.X.rst` file if fixing a bug or adding a new feature.

Ref: https://github.com/pandas-dev/pandas/pull/57971#discussion_r1539849666

ASVs against main:

```
| Change   | Before [b63ae8c7]    | After [46bbd5eb] <cln_argmin_argmax>   |   Ratio | Benchmark (Parameter)                                         |
|----------|----------------------|----------------------------------------|---------|---------------------------------------------------------------|
| -        | 184±1μs              | 138±3μs                                |    0.75 | series_methods.NanOps.time_func('argmax', 1000000, 'int32')   |
| -        | 14.4±0.1μs           | 10.5±0.2μs                             |    0.73 | series_methods.NanOps.time_func('argmax', 1000, 'float64')    |
| -        | 1.23±0.06ms          | 863±20μs                               |    0.7  | series_methods.NanOps.time_func('argmax', 1000000, 'float64') |
| -        | 77.9±0.6μs           | 40.8±0.7μs                             |    0.52 | series_methods.NanOps.time_func('argmax', 1000000, 'int8')    |
| -        | 7.42±0.2μs           | 2.44±0.06μs                            |    0.33 | series_methods.NanOps.time_func('argmax', 1000, 'int64')      |
| -        | 7.18±0.2μs           | 2.29±0.02μs                            |    0.32 | series_methods.NanOps.time_func('argmax', 1000, 'int32')      |
| -        | 7.25±0.1μs           | 2.29±0.03μs                            |    0.32 | series_methods.NanOps.time_func('argmax', 1000, 'int8')       |
```

ASVs against 2.2.x show no perf change.